+++ duckscript_sdk/src/sdk/std/fs/mod.rs
@@ -4,6 +4,7 @@ mod dirname;
 mod mkdir;
 mod print;
 mod read;
+mod rmdir;
 mod touch;
 mod write;
 
@@ -22,6 +23,7 @@ pub(crate) fn load(commands: &mut Commands, parent: &str) -> Result<(), ScriptEr
     commands.set(mkdir::create(&package))?;
     commands.set(print::create(&package))?;
     commands.set(read::create(&package))?;
+    commands.set(rmdir::create(&package))?;
     commands.set(touch::create(&package))?;
     commands.set(write::create(&package))?;